+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - add table `users_2fa_codes` to store 2FA codes
  - adjust procedure `add_table_column` and `drop_table_column` to add or drop a column from a table to work with special keywords used for table or column names
  - for `json` field type, if decode failed or still a string (likely double-encoded) then try to strip slashes and decode again
+ - add `email_2fa` and `email_2fa_subject` fields to `email` page
 
 ### Bugfix
  - properly delete jobs for action with enabled option `Clear Scheduled Jobs for This Action`. Before it deleted all jobs based on this action, now only jobs for the current user are deleted.